E-SQL/requirements.txt

argparse
python-dotenv==1.0.1
sqlglot==25.7.1
rank-bm25==0.2.2
nltk>=3.6.2
func_timeout==4.3.5
openai==1.37.1
numpy>=1.19.5
pandas>=1.1.5
sentencepiece==0.1.96
psycopg2-binary>=2.9.7


sqlparse==0.4.2
datasets>=1.11.0
pyarrow>=5.0.0


torch>=1.7.1
transformers>=4.9.2
huggingface-hub>=0.0.12
tokenizers>=0.10.3
accelerate>=0.12.0


scikit-learn>=0.24.2
scipy>=1.7.0
tqdm>=4.62.3
pyyaml>=5.4.1

requests>=2.26.0
click>=8.0.1
cffi>=1.14.6
portalocker>=2.3.2
